split out RemoveFromPicklistMenuItem so it can be used on the app side like AddToPicklistMenuItem
