+++ packages/components/releaseNotes/components.md
@@ -11,6 +11,7 @@ Components, models, actions, and utility functions for LabKey applications and p
     * move SamplesTabbedGridPanel from LKSM to ui-components
     * move CHECKED_OUT_BY_FIELD and INVENTORY_COLS constants to ui-components
     * Picklist implementation using SamplesTabbedGridPanel
+    * split out RemoveFromPicklistMenuItem so it can be used on the app side like AddToPicklistMenuItem
 
 ### version 2.91.0
 *Released*: 2 November 2021
